Refactor AST location helpers
